Refactored code - initial scope complete

+++ app/Http/Controllers/Admin/RoomsController.php
@@ -2,7 +2,7 @@
 
 namespace App\Http\Controllers\Admin;
 
-use App\Models\Admin\Room;
+use App\Models\Room;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Gate;
 use App\Http\Controllers\Controller;
@@ -18,7 +18,19 @@ class RoomsController extends Controller
      */
     public function index()
     {
-        $rooms = Room::all();
+        if (! Gate::allows('room_access')) {
+            return abort(401);
+        }
+
+
+        if (request('show_deleted') == 1) {
+            if (! Gate::allows('room_delete')) {
+                return abort(401);
+            }
+            $rooms = Room::onlyTrashed()->get();
+        } else {
+            $rooms = Room::all();
+        }
 
         return view('admin.rooms.index', compact('rooms'));
     }
@@ -30,6 +42,9 @@ class RoomsController extends Controller
      */
     public function create()
     {
+        if (! Gate::allows('room_create')) {
+            return abort(401);
+        }
         return view('admin.rooms.create');
     }
 
@@ -41,9 +56,14 @@ class RoomsController extends Controller
      */
     public function store(StoreRoomsRequest $request)
     {
+        if (! Gate::allows('room_create')) {
+            return abort(401);
+        }
         $room = Room::create($request->all());
 
-        return redirect()->route('rooms.index');
+
+
+        return redirect()->route('admin.rooms.index');
     }
 
 
@@ -55,6 +75,9 @@ class RoomsController extends Controller
      */
     public function edit($id)
     {
+        if (! Gate::allows('room_edit')) {
+            return abort(401);
+        }
         $room = Room::findOrFail($id);
 
         return view('admin.rooms.edit', compact('room'));
@@ -69,10 +92,15 @@ class RoomsController extends Controller
      */
     public function update(UpdateRoomsRequest $request, $id)
     {
+        if (! Gate::allows('room_edit')) {
+            return abort(401);
+        }
         $room = Room::findOrFail($id);
         $room->update($request->all());
-        
-        return redirect()->route('rooms.index');
+
+
+
+        return redirect()->route('admin.rooms.index');
     }
 
 
@@ -84,7 +112,10 @@ class RoomsController extends Controller
      */
     public function show($id)
     {
-        $bookings = \App\Models\Admin\Booking::where('room_id', $id)->get();
+        if (! Gate::allows('room_view')) {
+            return abort(401);
+        }
+        $bookings = \App\Booking::where('room_id', $id)->get();
 
         $room = Room::findOrFail($id);
 
@@ -100,9 +131,66 @@ class RoomsController extends Controller
      */
     public function destroy($id)
     {
+        if (! Gate::allows('room_delete')) {
+            return abort(401);
+        }
         $room = Room::findOrFail($id);
         $room->delete();
 
-        return redirect()->route('rooms.index');
+        return redirect()->route('admin.rooms.index');
+    }
+
+    /**
+     * Delete all selected Room at once.
+     *
+     * @param Request $request
+     */
+    public function massDestroy(Request $request)
+    {
+        if (! Gate::allows('room_delete')) {
+            return abort(401);
+        }
+        if ($request->input('ids')) {
+            $entries = Room::whereIn('id', $request->input('ids'))->get();
+
+            foreach ($entries as $entry) {
+                $entry->delete();
+            }
+        }
+    }
+
+
+    /**
+     * Restore Room from storage.
+     *
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function restore($id)
+    {
+        if (! Gate::allows('room_delete')) {
+            return abort(401);
+        }
+        $room = Room::onlyTrashed()->findOrFail($id);
+        $room->restore();
+
+        return redirect()->route('admin.rooms.index');
+    }
+
+    /**
+     * Permanently delete Room from storage.
+     *
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function perma_del($id)
+    {
+        if (! Gate::allows('room_delete')) {
+            return abort(401);
+        }
+        $room = Room::onlyTrashed()->findOrFail($id);
+        $room->forceDelete();
+
+        return redirect()->route('admin.rooms.index');
     }
 }
